Add Haskell Runtime interface
This commit does quite a number of things but the important part is that
it adds two functions that allow one to start and stop the Haskell
Runtime from one's code without needing to deal with all of the library
linking. The build script now goes to the ghc library directory where the
code is located and finds all of the .so files and links against them
to avoid dependency issues and making it easier to deal with other
Haskell code in the future. This means the user doesn't need to worry
about linking when it comes to getting the right Haskell libraries
involved beyond their own library they've created. They also just need
to use the library to start and stop the runtime without worrying about
getting the right C code types lined up just right.