Add configuration and min library for Haskell RTS
The build script didn't have the ability to compile the right files on
Mac or Windows because the dylib extension is different between the
three platforms. This commit changes that. It also filters the .so files
for the minimum amount of linking needed to get the library to run the
Haskell Runtime System. It also has the ability to choose between the
5 libraries based off of feature flags. Non-threaded is the default.