Structure Haskell modules properly
The library telling people to do imports with things like:

import Curryrs.Types

in Haskell. The problem is that you had to do:

import Types

This is not desired behavior and Types is pretty ambiguous. This
restructures the code to fit with the documentation for the new README
coming up and what code is coming out soon.