Add Haskell Runtime interface (#9)
* Add Haskell Runtime interface

This commit does quite a number of things but the important part is that
it adds two functions that allow one to start and stop the Haskell
Runtime from one's code without needing to deal with all of the library
linking. The build script now goes to the ghc library directory where the
code is located and finds all of the .so files and links against them
to avoid dependency issues and making it easier to deal with other
Haskell code in the future. This means the user doesn't need to worry
about linking when it comes to getting the right Haskell libraries
involved beyond their own library they've created. They also just need
to use the library to start and stop the runtime without worrying about
getting the right C code types lined up just right.

* Cleanup build.rs and comment it

* Add configuration and min library for Haskell RTS

The build script didn't have the ability to compile the right files on
Mac or Windows because the dylib extension is different between the
three platforms. This commit changes that. It also filters the .so files
for the minimum amount of linking needed to get the library to run the
Haskell Runtime System. It also has the ability to choose between the
5 libraries based off of feature flags. Non-threaded is the default.

* Fix typo in Cargo.toml

* Update .editorconfig to make Github readable

* Structure Haskell modules properly

The library telling people to do imports with things like:

import Curryrs.Types

in Haskell. The problem is that you had to do:

import Types

This is not desired behavior and Types is pretty ambiguous. This
restructures the code to fit with the documentation for the new README
coming up and what code is coming out soon.

* Update README.md to include HsRT instructions