Upgrade code and get it all working once again
This fixes year long issues in regards to linking. It's been frustrating
to say the least and was a big reason I let the code bit rot. Given
their is a pandemic goin on now and well I got bored I took the time to
fix all the code. This moves it into Rust 2018 edition works on a newer
version of GHC and overall reformats and cleans up the code with some
better libraries. This isn't the state the release will be in, but it's
just nice to have all the test suites working again and maybe now I can
make some actual progress on creating a good library.